standalone HGVertex & Region
